+++ checkpoint/orbax/checkpoint/experimental/v1/_src/layout/orbax_v0_layout.py
@@ -204,7 +204,10 @@ class OrbaxV0Layout(CheckpointLayout):
     if not await async_path.is_dir(path):
       raise NotADirectoryError(f"Checkpoint path {path} is not a directory.")
 
-    if await temporary_paths.is_path_temporary(path):
+    context = context_lib.get_context()
+    if await temporary_paths.is_path_temporary(
+        path, atomicity_options=context.atomicity.v0()
+    ):
       raise ValueError(f"Found incomplete checkpoint at {path}.")
 
     if await orbax_layout.has_checkpoint_metadata_file(path.parent):